Organize src directory
