+++ src/data/liesDictionary.ts
